# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  router.ErrorResponse:
    properties:
      error:
        type: string
      request_id:
        type: string
    type: object
...
  router.hostCommandRequest:
    properties:
      command:
        type: string
      environment:
        additionalProperties:
          type: string
        type: object
      shell:
        type: string
      timeout_seconds:
        type: integer
      working_directory:
        type: string
    required:
    - command
    type: object
  router.hostCommandResponse:
    properties:
      duration_ms:
        type: integer
      exit_code:
        type: integer
      stderr:
        type: string
      stdout:
        type: string
      timed_out:
        type: boolean
    type: object
...
info:
  contact: {}
  description: API documentation for the FeatherWings daemon.
  title: FeatherWings API
  version: "1.0"
paths:
...
  /api/system/terminal/exec:
    post:
      consumes:
      - application/json
      description: Runs a command on the host operating system using the configured
        shell and returns stdout/stderr.
      parameters:
      - description: Host command request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.hostCommandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.hostCommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "504":
          description: Command timed out
          schema:
            $ref: '#/definitions/router.hostCommandResponse'
      security:
      - NodeToken: []
      summary: Execute host command
      tags:
      - System
...
schemes:
- https
- http
securityDefinitions:
  NodeToken:
    description: 'Supply the node''s bearer token from `config.yml` using the `Authorization:
      Bearer <token>` header.'
    in: header
    name: Authorization
    type: apiKey
  ServerJWT:
    description: Signed JWTs issued by the Panel for server-scoped operations (uploads,
      downloads, websockets). Pass in the `token` query parameter.
    in: query
    name: token
    type: apiKey
swagger: "2.0"